Move `target_layers` to `GradientMetric`
This allows a more flexible specification of the gathered metrics and
allows it to use the metrics standalone.
